모드버스 서버 동작 잘 하는 듯 코일/홀드는

+++ ModBusQ/IModBusServer.cs
@@ -24,5 +24,5 @@ public interface IModBusServer : IDisposable
 	bool IsFunctionEnable(ModBusFunction function);
 
 	void WriteCoils(int address, params bool[] values);
-	void WriteHoldingRegisters(int address, params short[] values);
+	void WriteHoldingRegisters(int address, params int[] values);
 }
